Create `AssetData` concern
Prevent duplication of shared logic by making a new `AssetData` concern
that can be shared across the different types of asset data (`ImageData`
and `AttachmentData`). Can also eventually be used in all models that
implement similar logic (`ConsultationResponseFormData` and
`CallForEvidenceResponseFormData`).
